Create windows stubs for fonts and player
Former-commit-id: a722dbe1ed8ff77caa6e7d4f9830fb7606a3e7a8

+++ include/States/Menu.hpp
@@ -28,7 +28,7 @@ namespace SuperHaxagon {
 		int transitionFrame = 0;
 		int transitionDirection = 0;
 		int lastLevel = 0;
-		size_t level = 0;
+		int level = 0;
 	};
 }
